[refactor] 이미지 저장 버튼 변경
